% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_goal.R
\name{print_goal}
\alias{print_goal}
\title{Make a sorted frequency table for a factor}
\usage{
print_goal()
}
\arguments{
\item{x}{factor}
}
\value{
A tibble
}
\description{
Make a sorted frequency table for a factor
}
\examples{

}
